test(EuiDataGridCell): move EuiDataGrid column sizing jest tests to a separate file
